Added Blueprint for doctor
